% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{describe_table}
\alias{describe_table}
\title{Describe a data frame's columns and data types}
\usage{
describe_table(dataframe)
}
\arguments{
\item{dataframe}{A data frame}
}
\value{
A data frame with two columns: col containing the column names, and
  col_types containing the column data types
}
\description{
Describe a data frame's columns and data types
}
\examples{

}
