% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_zip_file}
\alias{get_zip_file}
\title{Download and extract files from a URL zipfile}
\usage{
get_zip_file(url, extractdir)
}
\arguments{
\item{url}{Address of zipfile}

\item{extractdir}{Folder to unzip to}
}
\value{
The contents of the zipfile are extracted into \code{extractdir}
}
\description{
Downloads a zipfile from a specific URL and extracts all of its contents into
a specified folder
}
