% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_airport_data.R
\name{read_airport_data}
\alias{read_airport_data}
\title{Read airport reference data}
\usage{
read_airport_data()
}
\value{
A data frame of airport data
}
\description{
Downloads the country information data from
the \url{www.uktradeinfo.com} website.
}
\examples{

}
