% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_country_data.R
\name{read_country_data}
\alias{read_country_data}
\title{Read country reference data}
\usage{
read_country_data()
}
\value{
A data frame of country data
}
\description{
Downloads the country information data from
the \url{www.uktradeinfo.com} website.
}
\examples{


}
