% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hmrc_data.R
\name{read_hmrc_data}
\alias{read_hmrc_data}
\title{Read an HMRC data file}
\usage{
read_hmrc_data(file, datatype)
}
\arguments{
\item{file}{An HMRC data file}

\item{datatype}{A data file type. One of "control", "exports",
"imports", "dispatches", "dispest", "arrivals", "arrest"}
}
\value{
A data frame of trade data
}
\description{
Uses the data specifications as detailed in
\url{https://www.uktradeinfo.com/Statistics/Pages/Supporting-material.aspx}
\tabular{ll}{
\strong{Type description} \tab \strong{Specification} \cr
control \tab SMKA12 \cr
exports \tab SMKE19 \cr
imports \tab SMKI19 \cr
arrivals/dispatches \tab SMKM46/SMKX46 \cr
arrest/dispest \tab SESM16/SESX16 estimates data
}
}
\examples{

}
