% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hmrc_traders.R
\name{read_hmrc_traders}
\alias{read_hmrc_traders}
\title{Read an HMRC traders file}
\usage{
read_hmrc_traders(file)
}
\arguments{
\item{file}{An HMRC tab delimited traders file}
}
\value{
A data frame of trader data in long form
}
\description{
Converts a standard HMRC exporter or importer file to long form for further
processing or analysis.
}
\examples{
data <- read_hmrc_traders("traderfile.tsv")
head(data)

}
