% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_icd_data.R
\name{read_icd_data}
\alias{read_icd_data}
\title{Read Inland Clearance Depot data}
\usage{
read_icd_data()
}
\value{
A data frame of Inland Clearance Depot data
}
\description{
Downloads information on Inland Clearance Depots and
Freezones from the \url{www.uktradeinfo.com} website.
}
\examples{

}
