% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_lcf_data.R
\name{read_lcf_data}
\alias{read_lcf_data}
\title{Read Local Clearance Freight locations data}
\usage{
read_lcf_data()
}
\value{
A data frame of Inland Clearance Depot data
}
\description{
Downloads information on NES Local Local Clearance Freight locations from the
\url{www.uktradeinfo.com} website.
}
\examples{

}
