% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_seaport_data.R
\name{read_seaport_data}
\alias{read_seaport_data}
\title{Read seaport reference data}
\usage{
read_seaport_data()
}
\value{
A data frame of seaport data
}
\description{
Downloads country information data from
the \url{www.uktradeinfo.com} website.
}
\examples{

}
