% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{safe_names}
\alias{safe_names}
\title{Sanitize table field names}
\usage{
safe_names(names)
}
\arguments{
\item{names}{A vector of strings}
}
\value{
A vector of strings with punctuation and spaces removed, and replaced
  with underscores
}
\description{
Removes odd characters from field names and makes them lower case. Converts
periods to underscores. Makes table field names safe to use in PostgreSQL
databases.
}
\examples{
x <- c("Isabella", "Andrew", "Graham", "Leigh")
y <- c(10, 25, 19, 36)
df <- data.frame( "First Name" = x, "High score! in 2018" = y)
colnames(df) <- safe_names(colnames(df))
}
