% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_data_arrdisest}
\alias{sample_data_arrdisest}
\title{EU import/export estimates data}
\format{A data frame with 100 rows and 4 variables:

\tabular{lll}{
\strong{Name} \tab \strong{Type} \tab \strong{Description}\cr
sitc_2          \tab character \tab x\cr
sitc_0          \tab character \tab x\cr
month_ind       \tab character \tab x\cr
estimated_value \tab double    \tab x
}}
\source{
\url{http://www.uktradeinfo.com/}
}
\usage{
sample_data_arrdisest
}
\description{
A dataset containing a random sample of 100 records from an HMRC arrival or
dispatch estimate download file (SESM16/SESX16).
}
\keyword{datasets}
