% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_data_traders}
\alias{sample_data_traders}
\title{HMRC trader data}
\format{A data frame with 100 rows and 31 variables:


\tabular{lll}{
\strong{Name}        \tab \strong{Type}      \tab \strong{Description}\cr
year        \tab double    \tab Year\cr
month       \tab double    \tab Month\cr
period      \tab integer   \tab Period in form YYYYMM\cr
nolines     \tab integer   \tab Number of lines for the company\cr
companyname \tab character \tab Name of company\cr
address1    \tab character \tab Address line 1\cr
address2    \tab character \tab Address line 2\cr
address3    \tab character \tab Address line 3\cr
address4    \tab character \tab Address line 4\cr
address5    \tab character \tab Address line 5\cr
postcode    \tab character \tab Postcode\cr
codenum     \tab character \tab Sequence number of commodity code\cr
comcode     \tab character \tab Commodity code
}}
\source{
\url{http://www.uktradeinfo.com/}
}
\usage{
sample_data_traders
}
\description{
A dataset containing a random sample of 100 records of HMRC importer/exporter
details data. The data is in the format returned by the
\code{\link{read_hmrc_traders}} function in the trade tools package. The
original files on the \url{www.uktradeinfo.com} site are named
Importers/Exporters.
}
\keyword{datasets}
