% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradetools.R
\docType{package}
\name{tradetools}
\alias{tradetools}
\alias{tradetools-package}
\title{tradetools: A package for working with HMRC overseas trade statistics}
\description{
The tradetools package contains a number of functions to download and process
HMRC trade data as published in \url{www.uktradeinfo.com}. There are
functions for importer and exporter data.
}
