% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hmrc_traders.R
\name{read_hmrc_traders}
\alias{read_hmrc_traders}
\title{Read an HMRC traders file}
\usage{
read_hmrc_traders(file)
}
\arguments{
\item{file}{An HMRC tab delimited traders file}
}
\description{
Read an HMRC traders file
}
