% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{safe_names}
\alias{safe_names}
\title{Sanitize table field names}
\usage{
safe_names(names)
}
\arguments{
\item{names}{}
}
\description{
Removes odd characters from field names and makes them lower case. Converts
periods to underscores. Makes table field names safe to use in PostgreSQL
databases.
}
