% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_data_arrdis}
\alias{sample_data_arrdis}
\title{EU import/export data}
\format{A data frame with 100 rows and 17 variables:
\describe{
  \item{comcode}{character}
  \item{record_type}{character}
  \item{cod_seq}{character}
  \item{cod_alpha}{character}
  \item{trade_ind}{character}
  \item{coo_seq}{character}
  \item{nature_of_transaction}{character}
  \item{mode_of_transport}{character}
  \item{period_reference}{character}
  \item{suite_indicator}{character}
  \item{sitc}{character}
  \item{ip_comcode}{character}
  \item{stat_value}{double}
  \item{nett_mass}{double}
  \item{supp_unit}{double}
}}
\source{
\url{http://www.uktradeinfo.com/}
}
\usage{
sample_data_arrdis
}
\description{
A dataset containing a random sample of 100 records from an HMRC arrival or
dispatch download file (SMKM46/SMKX46).
}
\keyword{datasets}
