% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_data_arrdisest}
\alias{sample_data_arrdisest}
\title{EU import/export estimates data}
\format{A data frame with 100 rows and 4 variables:
\describe{
  \item{sitc_2}{character}
  \item{sitc_0}{character}
  \item{month_ind}{character}
  \item{estimated_value}{double}
}}
\source{
\url{http://www.uktradeinfo.com/}
}
\usage{
sample_data_arrdisest
}
\description{
A dataset containing a random sample of 100 records from an HMRC arrival or
dispatch estimate download file (SESM16/SESX16).
}
\keyword{datasets}
