% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_data_control}
\alias{sample_data_control}
\title{HMRC control file data}
\format{A data frame with 100 rows and 31 variables:
\describe{
  \item{comcode}{character}
  \item{intra_extra_ind}{character}
  \item{intra_mm_on}{character}
  \item{intra_yy_on}{character}
  \item{intra_mm_off}{character}
  \item{intra_yy_off}{character}
  \item{extra_mm_on}{character}
  \item{extra_yy_on}{character}
  \item{extra_mm_off}{character}
  \item{extra_yy_off}{character}
  \item{non_trade_id}{character}
  \item{sitc_no}{character}
  \item{sitc_ind}{character}
  \item{sitc_conv_a}{character}
  \item{sitc_conv_b}{character}
  \item{cn_q2}{character}
  \item{supp_arrivals}{character}
  \item{supp_dispatches}{character}
  \item{supp_imports}{character}
  \item{supp_exports}{character}
  \item{sub_group_arr}{character}
  \item{item_arr}{character}
  \item{sub_group_disp}{character}
  \item{item_disp}{character}
  \item{sub_group_imp}{character}
  \item{item_imp}{character}
  \item{sub_group_exp}{character}
  \item{item_exp}{character}
  \item{qty1_alpha}{character}
  \item{qty2_alpha}{character}
  \item{commodity_alpha}{character}
}}
\source{
\url{http://www.uktradeinfo.com/}
}
\usage{
sample_data_control
}
\description{
A dataset containing a random sample of 100 records from an HMRC control file
(SMKA12).
}
\keyword{datasets}
