% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_data_exports}
\alias{sample_data_exports}
\title{Non-EU exports data}
\format{A data frame with 100 rows and 23 variables:
\describe{
  \item{comcode}{character}
  \item{sitc}{character}
  \item{record_type}{character}
  \item{cod_seq}{character}
  \item{cod_alpha}{character}
  \item{account_mm}{character}
  \item{account_ccyy}{character}
  \item{port_seq}{character}
  \item{port_alpha}{character}
  \item{flag_seq}{character}
  \item{flag_alpha}{character}
  \item{trade_ind}{character}
  \item{container}{character}
  \item{mode_of_transport}{character}
  \item{inland_mot}{character}
  \item{golo_seq}{character}
  \item{golo_alpha}{character}
  \item{suite_indicator}{character}
  \item{procedure_code}{character}
  \item{value}{double}
  \item{quantity_1}{double}
  \item{quantitiy_2}{double}
  \item{industrial_plant_comcode}{character}
}}
\source{
\url{http://www.uktradeinfo.com/}
}
\usage{
sample_data_exports
}
\description{
A dataset containing a ramdom sample of 100 records from an HMRC export
download file (SMKE19).
}
\keyword{datasets}
