% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{unzip_and_cleanup}
\alias{unzip_and_cleanup}
\title{Tidy up HMRC data files}
\usage{
unzip_and_cleanup(extractdir)
}
\arguments{
\item{extractdir}{folder containing data files}
}
\value{
The contents of \code{extractdir} are renamed as appropriate.
Any .zip files in the folder are deleted.
}
\description{
Unzips all zipfiles in a folder and removes the zipfiles. Tidies up
the filenames of the downloads.
}
\details{
All files are set to lower case. HMRC trader details files have
..txt suffixes changed to .tsv. Inconsistent filenames in
February 2009 and May/June 2014 data are fixed.
}
\examples{
unzip_and_cleanup("~/temp")


}
